refactor, extract api_server mod.rs to mutiple files base on functionability
