+++ rust/sglang-server/src/api_server/common.rs
@@ -1,0 +1,263 @@
+//! Common control-plane endpoints — `/server_info`, `/get_model_info`
+//! (+ `/model_info` alias), plus the control-request submission path
+//! (`submit` / `await_control_result`). Data-plane endpoints (incl. `/health*`,
+//! which round-trips a generate probe) live in the sibling `sglang` (native
+//! protocol) and `openai` modules; the shared `AppState` lives in the parent
+//! `api_server` module.
+
+use axum::{
+    Router,
+    extract::State,
+    http::StatusCode,
+    response::{IntoResponse, Response},
+    routing::get,
+};
+use tokio::sync::mpsc;
+
+use super::AppState;
+use crate::fsm::RequestState;
+use crate::ids::RidHash;
+use crate::message::{ControlRequest, EgressItem, EgressSink, Request, RequestKind};
+use crate::runtime::ServerArgs;
+use crate::runtime::channels::TmEvent;
+
+/// The routes this module owns, mounted by `api_server::serve`.
+pub(super) fn routes() -> Router<AppState> {
+    Router::new()
+        // Control-plane: reuses the ingress FSM (no tokenization), returns one
+        // non-streamed JSON result. Adding one = a route line + its struct tag.
+        .route("/server_info", get(server_info))
+        // Static config, no scheduler round-trip. `/get_model_info` (+ `/model_info`
+        // alias) is what the SGLang lang backend (`RuntimeEndpoint`, gsm8k/eval)
+        // calls at startup.
+        .route("/get_model_info", get(model_info))
+        .route("/model_info", get(model_info))
+}
+
+/// Submit one control request into the ingress pipeline (always a rust-minted
+/// rid — control responses are routed by it, so no client-supplied form
+/// exists); returns the rid, its hashed routing key, and the egress receiver.
+async fn submit(
+    state: &AppState,
+    tag: &'static str,
+) -> Result<(RidHash, String, mpsc::Receiver<EgressItem>), ()> {
+    let rid = crate::ids::new_rid();
+    let id = RidHash::from_rid(&rid);
+    // Async-aware send so a full TM inbox yields (backpressure) instead of parking
+    // a thread; Err only when the inbox is closed (shutdown).
+    let (tx, rx) = mpsc::channel::<EgressItem>(state.egress_buf);
+    let request = Request {
+        rid_hash: id,
+        rid: rid.clone(),
+        state: RequestState::Received,
+        sink: EgressSink::Local(tx),
+        kind: RequestKind::Control(ControlRequest { tag }),
+    };
+    match state.senders.tm.send_async(TmEvent::Ingress(request)).await {
+        Ok(()) => Ok((id, rid, rx)),
+        Err(_) => {
+            tracing::error!("tm inbox closed; request dropped");
+            Err(())
+        }
+    }
+}
+
+/// Submit a `Control(tag)` through the ingress FSM (no tokenization) and await the
+/// scheduler's single msgpack result (a `structs.asdict` named map). Returns the
+/// raw bytes, or an error `Response` to return as-is.
+async fn await_control_result(
+    state: &AppState,
+    tag: &'static str,
+) -> Result<bytes::Bytes, Response> {
+    let (_id, _rid, mut rx) = submit(state, tag)
+        .await
+        .map_err(|()| (StatusCode::SERVICE_UNAVAILABLE, "service unavailable").into_response())?;
+    match rx.recv().await {
+        Some(EgressItem::Control(bytes)) => Ok(bytes),
+        Some(EgressItem::Error(e)) => {
+            let code =
+                StatusCode::from_u16(e.http_status()).unwrap_or(StatusCode::INTERNAL_SERVER_ERROR);
+            Err((code, e.to_string()).into_response())
+        }
+        // A control request never receives generation frames.
+        Some(EgressItem::Frame(_)) | Some(EgressItem::Done(_)) => Err((
+            StatusCode::INTERNAL_SERVER_ERROR,
+            "unexpected generation output for control request",
+        )
+            .into_response()),
+        None => Err((StatusCode::from_u16(499).unwrap(), "request aborted").into_response()),
+    }
+}
+
+/// Generic control endpoint: the scheduler's response straight to JSON (`tag` =
+/// request-struct name). For control endpoints whose response needs no shaping.
+#[allow(dead_code)] // first non-/server_info control endpoint will use this
+async fn control(State(state): State<AppState>, tag: &'static str) -> Response {
+    match await_control_result(&state, tag).await {
+        Ok(bytes) => match msgpack_to_json(&bytes) {
+            Ok(json) => {
+                (StatusCode::OK, [("content-type", "application/json")], json).into_response()
+            }
+            Err(e) => {
+                tracing::error!(error = %e, "control: msgpack→json failed");
+                (StatusCode::INTERNAL_SERVER_ERROR, "bad control response").into_response()
+            }
+        },
+        Err(resp) => resp,
+    }
+}
+
+/// Convert a msgpack control response (the scheduler's native ring format) into
+/// JSON bytes for the HTTP client.
+fn msgpack_to_json(bytes: &[u8]) -> Result<Vec<u8>, String> {
+    let val = rmpv::decode::read_value(&mut &*bytes).map_err(|e| e.to_string())?;
+    serde_json::to_vec(&val).map_err(|e| e.to_string())
+}
+
+/// `GET /get_model_info` (+ `/model_info` alias) — static model metadata from
+/// `server_args` (no scheduler round-trip); `is_generation` always true.
+async fn model_info(State(state): State<AppState>) -> Response {
+    let sa = &state.server_args;
+    let body = serde_json::json!({
+        "model_path": sa.model_path(),
+        "tokenizer_path": sa.tokenizer_path(),
+        "is_generation": true,
+        "preferred_sampling_params": serde_json::Value::Null,
+        "weight_version": serde_json::Value::Null,
+    });
+    (
+        StatusCode::OK,
+        [("content-type", "application/json")],
+        serde_json::to_vec(&body).unwrap_or_default(),
+    )
+        .into_response()
+}
+
+/// `GET /server_info` — surface only an allowlist ([`INTERNAL_STATE_ALLOWLIST`] +
+/// curated [`ServerArgs`] accessors), never the raw server-args dump (embeds
+/// `api_key`/`admin_api_key`; see [`shape_server_info`]).
+///
+/// TODO(server_info): Python also includes `kv_events`; add once plumbed.
+async fn server_info(State(state): State<AppState>) -> Response {
+    let bytes = match await_control_result(&state, "GetInternalStateReq").await {
+        Ok(b) => b,
+        Err(resp) => return resp,
+    };
+    match shape_server_info(&bytes, &state.server_args) {
+        Ok(json) => (StatusCode::OK, [("content-type", "application/json")], json).into_response(),
+        Err(e) => {
+            tracing::error!(error = %e, "server_info: shaping failed");
+            (
+                StatusCode::INTERNAL_SERVER_ERROR,
+                "bad server_info response",
+            )
+                .into_response()
+        }
+    }
+}
+
+/// Runtime-metric keys `get_internal_state` adds atop the server-args dump. We copy
+/// ONLY these out of `internal_state` (an allowlist), so the co-mingled
+/// `api_key`/`admin_api_key` can never reach the response.
+const INTERNAL_STATE_ALLOWLIST: &[&str] = &[
+    "last_gen_throughput",
+    "memory_usage",
+    "effective_max_running_requests_per_dp",
+    "avg_spec_accept_length",
+    "step_time_dict",
+];
+
+fn shape_server_info(msgpack: &[u8], server_args: &ServerArgs) -> Result<Vec<u8>, String> {
+    // GetInternalStateReqOutput asdict → `{ "internal_state": { server-args dump +
+    // metrics }, ... }`. Pull that inner map out (it is NOT safe to expose whole).
+    let mut obj: serde_json::Map<String, serde_json::Value> =
+        rmp_serde::from_slice(msgpack).map_err(|e| e.to_string())?;
+    let internal = match obj.remove("internal_state") {
+        Some(serde_json::Value::Object(m)) => m,
+        _ => serde_json::Map::new(),
+    };
+
+    // Copy only the allowlisted runtime metrics — never the raw server-args dump.
+    let mut state_out = serde_json::Map::new();
+    for &k in INTERNAL_STATE_ALLOWLIST {
+        match internal.get(k) {
+            Some(v) if !v.is_null() => {
+                state_out.insert(k.to_string(), v.clone());
+            }
+            _ => {}
+        }
+    }
+
+    // Top-level non-secret config from typed accessors (structurally can't surface
+    // a key field, unlike the raw dump).
+    let response = serde_json::json!({
+        "model_path": server_args.model_path(),
+        "served_model_name": server_args.served_model_name(),
+        "tokenizer_path": server_args.tokenizer_path(),
+        "max_context_length": server_args.context_len(),
+        "max_total_num_tokens": server_args.max_total_num_tokens(),
+        "version": server_args.version(),
+        "internal_states": [serde_json::Value::Object(state_out)],
+    });
+    serde_json::to_vec(&response).map_err(|e| e.to_string())
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    /// The scheduler's `internal_state` embeds the full server-args dump (incl.
+    /// `api_key`/`admin_api_key`). `/server_info` must surface only the allowlisted
+    /// runtime metrics + curated config — never the secrets — and must not re-nest
+    /// the dump under `internal_states[].internal_state`.
+    #[test]
+    fn shape_server_info_excludes_secrets_and_dump() {
+        // GetInternalStateReqOutput.asdict → { "internal_state": { …dump+metrics… } }.
+        let internal = rmpv::Value::Map(vec![
+            (
+                rmpv::Value::from("api_key"),
+                rmpv::Value::from("secret-token"),
+            ),
+            (
+                rmpv::Value::from("admin_api_key"),
+                rmpv::Value::from("admin-token"),
+            ),
+            (rmpv::Value::from("model_path"), rmpv::Value::from("/m")),
+            (
+                rmpv::Value::from("last_gen_throughput"),
+                rmpv::Value::from(1.5),
+            ),
+            (
+                rmpv::Value::from("effective_max_running_requests_per_dp"),
+                rmpv::Value::from(32),
+            ),
+        ]);
+        let outer = rmpv::Value::Map(vec![(rmpv::Value::from("internal_state"), internal)]);
+        let mut msgpack = Vec::new();
+        rmpv::encode::write_value(&mut msgpack, &outer).unwrap();
+
+        let sa =
+            ServerArgs::from_json(r#"{"model_path": "/m", "api_key": "secret-token"}"#).unwrap();
+        let out = shape_server_info(&msgpack, &sa).unwrap();
+        let text = String::from_utf8(out.clone()).unwrap();
+        // No secret leaks anywhere in the serialized response.
+        assert!(!text.contains("secret-token"), "api_key leaked: {text}");
+        assert!(
+            !text.contains("admin-token"),
+            "admin_api_key leaked: {text}"
+        );
+
+        let v: serde_json::Value = serde_json::from_slice(&out).unwrap();
+        // Allowlisted metric surfaced; the whole dump did not.
+        let state0 = &v["internal_states"][0];
+        assert_eq!(state0["last_gen_throughput"], 1.5);
+        assert_eq!(state0["effective_max_running_requests_per_dp"], 32);
+        assert!(
+            state0.get("internal_state").is_none(),
+            "must not re-nest the dump under internal_state"
+        );
+        assert!(state0.get("api_key").is_none());
+        // Curated top-level config comes from typed accessors, not the dump.
+        assert_eq!(v["model_path"], "/m");
+    }
+}
